+++ CHANGES.md
@@ -1,6 +1,9 @@
 Unreleased
 ----------
 
+- Don't generate the dune-site build part when not needed (#4861, @bobot,
+  review @kit-ty-kate)
+
 - Fixes `opam` META file requires entry for private libs (#4841, fixes #4839, @toots)
 
 - Fixes `dune exec` not adding .exe on Windows (#4371, fixes #3322, @MisterDA)